# openapi.yaml
openapi: 3.1.0
info:
  title: DM Agent API
  version: "1.2.0"

servers:
  - url: https://ubiquitous-space-giggle-r49rgrw5g6qwhv9p-8000.app.github.dev

ccomponents:
  schemas:
    RollReq:
      type: object
      required: [expr]
      properties:
        expr: { type: string }
        session_id: { type: string }
        label: { type: string }

    SkillCheckReq:
      type: object
      required: [bonus, dc]
      properties:
        bonus: { type: integer }
        dc: { type: integer }
        mode:
          type: string
          enum: [normal, adv, dis]
          default: normal
        session_id: { type: string }
        label: { type: string }

    AttackReq:
      type: object
      required: [attack_bonus, target_ac]
      properties:
        attack_bonus: { type: integer }
        target_ac: { type: integer }
        mode:
          type: string
          enum: [normal, adv, dis]
          default: normal
        session_id: { type: string }
        label: { type: string }
        damage_expr: { type: string }

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY

security:
  - ApiKeyAuth: []

paths:
  /health:
    get:
      operationId: health
      summary: Health check
      security: []
      responses:
        "200":
          description: OK

  /config:
    get:
      operationId: config
      summary: Runtime config (safe)
      security: []
      responses:
        "200":
          description: OK

  /turn:
    post:
      operationId: turn
      summary: Run one DM agent turn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, text]
              properties:
                session_id:
                  type: string
                text:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [session_id, output]
                properties:
                  session_id:
                    type: string
                  output:
                    type: string

  /session/reset:
    post:
      operationId: sessionReset
      summary: Reset a session
      parameters:
        - name: session_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK

  /session/{session_id}:
    get:
      operationId: sessionDump
      summary: Dump stored session history
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  
    /roll:
    post:
      operationId: roll
      summary: Roll a dice expression (e.g., 2d6+3)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RollReq"
      responses:
        "200":
          description: OK

  /skill_check:
    post:
      operationId: skillCheck
      summary: Skill/ability check with bonus vs DC (supports adv/dis)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillCheckReq"
      responses:
        "200":
          description: OK

  /attack:
    post:
      operationId: attack
      summary: Attack roll vs target AC (supports adv/dis); optional damage roll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttackReq"
      responses:
        "200":
          description: OK